Refactor: moved to SCSS, cleaned up layout and styling
