--- src/App.css
@@ -1,86 +1,0 @@
-:root {
-  --bg: #f7f9fc;
-  --accent: #4a90e2;
-  --text: #333;
-  --done: #ccc;
-}
-
-body {
-  margin: 0;
-  font-family: 'Inter', sans-serif;
-  background-color: var(--bg);
-  color: var(--text);
-}
-
-#root {
-  max-width: 600px;
-  margin: 3rem auto;
-  padding: 2rem;
-  background: #fff;
-  border-radius: 12px;
-  box-shadow: 0 4px 24px rgba(0, 0, 0, 0.06);
-  text-align: left;
-}
-
-h1 {
-  font-size: 1.8rem;
-  margin-bottom: 1.5rem;
-}
-
-form {
-  display: flex;
-  gap: 0.5rem;
-  margin-bottom: 1.5rem;
-}
-
-input[type="text"] {
-  flex: 1;
-  padding: 0.75rem;
-  font-size: 1rem;
-  border: 1px solid #ddd;
-  border-radius: 8px;
-}
-
-button {
-  padding: 0.75rem 1rem;
-  font-size: 1rem;
-  background-color: var(--accent);
-  color: white;
-  border: none;
-  border-radius: 8px;
-  cursor: pointer;
-  transition: background-color 0.2s ease;
-}
-button:hover {
-  background-color: #3b78c7;
-}
-
-ul {
-  list-style: none;
-  padding: 0;
-}
-
-li {
-  display: flex;
-  align-items: center;
-  padding: 0.5rem 0;
-  border-bottom: 1px solid #eee;
-}
-
-li:last-child {
-  border-bottom: none;
-}
-
-input[type="checkbox"] {
-  margin-right: 0.75rem;
-  transform: scale(1.2);
-}
-
-li span {
-  flex: 1;
-}
-
-li span.done {
-  text-decoration: line-through;
-  color: var(--done);
-}
