+++ src/components/TaskInput/TaskInput.jsx
@@ -1,4 +1,5 @@
 import { useState } from 'react'
+import styles from './TaskInput.module.scss'
 
 export default function TaskInput({ onAdd }) {
     const [value, setValue] = useState('')
@@ -12,15 +13,15 @@ export default function TaskInput({ onAdd }) {
     }
 
     return (
-        <form onSubmit={handleSubmit} style={{ marginBottom: '1rem' }}>
+        <form onSubmit={handleSubmit} className={styles.taskForm}>
             <input
                 type="text"
                 value={value}
                 onChange={(e) => setValue(e.target.value)}
                 placeholder="Dodaj zadanie..."
-                style={{ padding: '0.5rem', width: '300px' }}
+                className={styles.taskInput}
             />
-            <button type="submit" style={{ marginLeft: '0.5rem' }}>Dodaj</button>
+            <button type="submit" className={styles.taskButton}>Dodaj</button>
         </form>
     )
 }
