+++ src/components/TaskList/TaskList.module.scss
@@ -1,0 +1,43 @@
+.taskItem {
+  display: flex;
+  align-items: center;
+  margin-bottom: 0.5rem;
+}
+
+.taskLabel {
+  display: flex;
+  align-items: center;
+  flex: 1;
+}
+
+.taskText {
+  margin-left: 0.5rem;
+  flex: 1;
+}
+
+.done {
+  text-decoration: line-through;
+  color: #999;
+}
+
+.editInput {
+  margin-left: 0.5rem;
+  flex: 1;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  padding: 0.25rem;
+}
+
+.editButton,
+.deleteButton {
+  margin-left: 0.5rem;
+  background: none;
+  border: none;
+  color: #555;
+  cursor: pointer;
+  font-size: 1rem;
+}
+
+.deleteButton {
+  color: #c00;
+}
