+++ src/index.scss
@@ -1,0 +1,7 @@
+/* index.scss – minimal reset */
+
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
