+++ src/main.jsx
@@ -1,8 +1,8 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
+import './index.scss'
 import App from './App.jsx'
-import './App.css'
+import './App.scss'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
